requirements.txt
Django==4.2.28
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.2.1
